// src/app/modules/racecard/models/horse-previous-run.ts
export class HorsePreviousRun {
    raceCardHorseHistoryId: number;
    sourceRaceId: number;
    raceDate: Date;
    raceTime: string;
    raceName: string;
    raceClass: number;
    going: string;
    distance: string;
    novice: boolean;
    maiden: boolean;
    handicap: boolean;
    hurdle: boolean;
    chase: boolean;
    nHF: boolean;
    raceType: string;
    aW: boolean;
    grade: number;
    irish: boolean;
    irishGrade: string;
    prizeMoney?: number;
    horseName: string;
    odds: string;
    position: number;
    placed?: boolean;
    failedToFinishReason: string;
    distanceBeaten: string;
    age: number;
    comment: string;
    trainer: string;
    jockey: string;
    jockeyClaim: number;
    headGear: string;
    officialRating: number;
    tS?: number;
    rPR?: number;
    weight: string;
    paceValue?: number;
    decimalOdds?: number;
    profitLoss?: number;
    course: string;
}